Modified methods temperatureConversion(), also type

+++ src/main/java/com/gmail/alinakotova102/task22/Converter.java
@@ -3,6 +3,6 @@ package com.gmail.alinakotova102.task22;
 public abstract class Converter {
     protected final float T0 = 273.15F;
 
-    abstract double temperatureConversion(int degreeCelsius);
+    abstract double temperatureConversion(double degreeCelsius);
 }
