+++ src/main/java/com/gmail/alinakotova102/task22/Fahrenheit.java
@@ -1,10 +1,8 @@
 package com.gmail.alinakotova102.task22;
 
-import org.junit.Test;
-
 public class Fahrenheit extends Converter {
     @Override
-    public double temperatureConversion(int degreeCelsius) {
-        return degreeCelsius * 1.8 + 32;
+    public double temperatureConversion(double degreeCelsius) {
+        return (degreeCelsius * 1.8) + 32;
     }
 }
