+++ src/main/java/com/gmail/alinakotova102/task22/Kelvin.java
@@ -3,7 +3,7 @@ package com.gmail.alinakotova102.task22;
 public class Kelvin extends Converter {
 
     @Override
-    public double temperatureConversion(int degreeCelsius) {
+    public double temperatureConversion(double degreeCelsius) {
         return degreeCelsius + T0;
     }
 }
